LPS-58563 Move for consistency
